IMPOSSIBLE: definition: goto definition from popup
The server needs positional arguments of the argument we want to go to. This is impossible
because the popup window is not saved in any file. Thus, the only option would be to know
where the desired word is in any file. The keyword does not have to be located in current
buffer.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>