parser: create parser module for free use
The parser module makes an easy way of usage of the parser
implementation. The only thing that needs to be done is to supply the
text to be parsed.

The output is then the parsed text and highlighting table which needs to
be applied once the text is displayed in a buffer.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>