fix(#34): parser: return empty ParserOutput on empty input
If the input to the parse function is nil / emtpy string / emtpy table,
an emtpy ParserOutput object will be returned. Meaning the text and the
highlighting fields will be set to empty tables.

To check if the returned object is emtpy you can do the following
local out = parser.parse(nil)
vim.print(out == out:emtpy() and "Equal" or "Different")

The out:empty() creates a new empty object which can be used for
comparison. The __eq method is the only comparison method implemented.

Signed-off-by: Filip Lobpreis <f.lobpreis@gmail.com>